+++ dgisim/src/status/status.py
@@ -122,6 +122,7 @@ __all__ = [
     "MushroomPizzaStatus",
     "NorthernSmokedChickenStatus",
     "SatiatedStatus",
+    "TandooriRoastChickenStatus",
     "UnmovableMountainStatus",
 
     # character specific status
@@ -2281,6 +2282,41 @@ class SatiatedStatus(CharacterStatus):
     REACTABLE_SIGNALS: ClassVar[frozenset[TriggeringSignal]] = frozenset((
         TriggeringSignal.ROUND_END,
     ))
+
+    @override
+    def _react_to_signal(
+            self, game_state: GameState, source: StaticTarget, signal: TriggeringSignal
+    ) -> tuple[list[eft.Effect], None | Self]:
+        if signal is TriggeringSignal.ROUND_END:
+            return [], None
+        return [], self  # pragma: no cover
+
+
+@dataclass(frozen=True, kw_only=True)
+class TandooriRoastChickenStatus(CharacterStatus):
+    DMG_BOOST: ClassVar[int] = 2
+    REACTABLE_SIGNALS: ClassVar[frozenset[TriggeringSignal]] = frozenset((
+        TriggeringSignal.ROUND_END,
+    ))
+
+    @override
+    def _preprocess(
+            self,
+            game_state: GameState,
+            status_source: StaticTarget,
+            item: PreprocessableEvent,
+            signal: Preprocessables,
+    ) -> tuple[PreprocessableEvent, None | Self]:
+        if signal is Preprocessables.DMG_AMOUNT_PLUS:
+            assert isinstance(item, DmgPEvent)
+            dmg = item.dmg
+            if (
+                dmg.source == status_source
+                and dmg.damage_type.direct_elemental_skill()
+                and dmg.element is not Element.PIERCING
+            ):
+                return item.delta_damage(self.DMG_BOOST), None
+        return super()._preprocess(game_state, status_source, item, signal)
 
     @override
     def _react_to_signal(